BTAPIClient - make methods that accept http type param public for use in Drop-in
